embryJayMilestone3
For the sake of time during playtesting, the first two stages do not last very long. This will change for the completed version. Also, the background is only like that because the art for it doesn't loop, which will also be changed for the final product.
(Amended in order to ensure that the score does not keep going up/the lives do not keep going down once the player has lost.)